Split SQLBindParameter C-to-VARCHAR tests into separate files
Restructure the c_to_varchar tests from sql_bind_parameter.cpp into
dedicated files following the file-per-conversion pattern used by
sql-to-c tests (e.g., real_conversion_to_c_char.cpp).

New files:
- c_integer_to_varchar.cpp (18 tests: integer types + boundary values)
- c_real_to_varchar.cpp (9 tests: double/float + NaN/Inf/-0 edge cases)
- c_bit_to_varchar.cpp (3 tests: true/false/clamping)
- c_char_to_varchar.cpp (6 tests: default + string SQL type variants)
- c_temporal_to_varchar.cpp (4 tests: timestamp/date/time)
- c_numeric_to_varchar.cpp (3 tests: integer/scale/negative scale)
- c_binary_to_varchar.cpp (1 test: hex encoding)

Each cpp file has a corresponding .feature file. The original
sql_bind_parameter.cpp retains error codes, API behavior, and
descriptor integration tests (30 tests).

Addresses Piotr's review comment on PR #748.
SNOW-3031761

Made-with: Cursor
