+++ odbc_tests/tests/e2e/CMakeLists.txt
@@ -18,6 +18,13 @@ add_odbc_test(e2e_query_basic_execute_query query/basic_execute_query.cpp)
 add_odbc_test(e2e_query_large_result_set query/large_result_set.cpp)
 add_odbc_test(e2e_query_sql_bind_col query/sql_bind_col.cpp)
 add_odbc_test(e2e_query_sql_bind_parameter query/sql_bind_parameter.cpp)
+add_odbc_test(e2e_query_c_integer_to_varchar query/c_integer_to_varchar.cpp)
+add_odbc_test(e2e_query_c_real_to_varchar query/c_real_to_varchar.cpp)
+add_odbc_test(e2e_query_c_bit_to_varchar query/c_bit_to_varchar.cpp)
+add_odbc_test(e2e_query_c_char_to_varchar query/c_char_to_varchar.cpp)
+add_odbc_test(e2e_query_c_temporal_to_varchar query/c_temporal_to_varchar.cpp)
+add_odbc_test(e2e_query_c_numeric_to_varchar query/c_numeric_to_varchar.cpp)
+add_odbc_test(e2e_query_c_binary_to_varchar query/c_binary_to_varchar.cpp)
 add_odbc_test(e2e_query_sql_describe_col query/sql_describe_col.cpp)
 add_odbc_test(e2e_query_sql_fetch query/sql_fetch.cpp)
 add_odbc_test(e2e_query_sql_get_data query/sql_get_data.cpp)
