+++ odbc_tests/tests/e2e/query/sql_bind_parameter.cpp
@@ -1,8 +1,4 @@
-#include <algorithm>
-#include <cctype>
-#include <climits>
 #include <cstring>
-#include <limits>
 #include <string>
 
 #include <catch2/catch_test_macros.hpp>
@@ -13,9 +9,6 @@
 #include "get_data.hpp"
 #include "odbc_cast.hpp"
 #include "odbc_matchers.hpp"
-
-// TODO(SNOW-3031761): Restructure binding tests to follow the same file-per-conversion
-// pattern used by the sql-to-c tests (e.g., real_conversion_to_c_char.cpp) for readability.
 
 // =============================================================================
 // ODBC SQLBindParameter spec-compliance tests
@@ -781,830 +774,4 @@ TEST_CASE("should report APD count zero when no parameters are bound.", "[query]
 
   // Then count should be 0
   CHECK(count == 0);
-}
-
-// =============================================================================
-// C type → VARCHAR conversions
-// =============================================================================
-
-TEST_CASE("should bind SQL_C_LONG to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_LONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind negative SQL_C_SLONG to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = -999;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "-999");
-}
-
-TEST_CASE("should bind SQL_C_ULONG to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLUINTEGER param = 4000000000U;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_ULONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "4000000000");
-}
-
-TEST_CASE("should bind SQL_C_SSHORT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLSMALLINT param = -7;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SSHORT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "-7");
-}
-
-TEST_CASE("should bind SQL_C_SHORT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLSMALLINT param = -7;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SHORT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "-7");
-}
-
-TEST_CASE("should bind SQL_C_USHORT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLUSMALLINT param = 65535;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_USHORT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "65535");
-}
-
-TEST_CASE("should bind SQL_C_SBIGINT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLBIGINT param = 9999999999LL;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SBIGINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "9999999999");
-}
-
-TEST_CASE("should bind SQL_C_UBIGINT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLUBIGINT param = 1000000000000ULL;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_UBIGINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "1000000000000");
-}
-
-TEST_CASE("should bind SQL_C_STINYINT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLSCHAR param = -128;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_STINYINT, SQL_VARCHAR, 100, 0, &param, 0,
-                                   &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "-128");
-}
-
-TEST_CASE("should bind SQL_C_TINYINT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLSCHAR param = -128;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_TINYINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "-128");
-}
-
-TEST_CASE("should bind SQL_C_UTINYINT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLCHAR param = 255;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_UTINYINT, SQL_VARCHAR, 100, 0, &param, 0,
-                                   &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "255");
-}
-
-TEST_CASE("should bind SQL_C_DOUBLE to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLDOUBLE param = 3.14;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DOUBLE, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "3.14");
-}
-
-TEST_CASE("should bind SQL_C_FLOAT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLREAL param = 1.5f;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_FLOAT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "1.5");
-}
-
-TEST_CASE("should bind SQL_C_BIT true to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLCHAR param = 1;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_BIT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "1");
-}
-
-TEST_CASE("should bind SQL_C_BIT false to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLCHAR param = 0;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_BIT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "0");
-}
-
-// =============================================================================
-// Boundary values, edge cases, and negative tests
-// =============================================================================
-
-TEST_CASE("should bind SQL_C_SLONG INT_MIN to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = INT_MIN;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == std::to_string(INT_MIN));
-}
-
-TEST_CASE("should bind SQL_C_SLONG INT_MAX to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = INT_MAX;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == std::to_string(INT_MAX));
-}
-
-TEST_CASE("should bind SQL_C_SLONG zero to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 0;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "0");
-}
-
-TEST_CASE("should bind SQL_C_SBIGINT LLONG_MIN to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLBIGINT param = LLONG_MIN;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SBIGINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == std::to_string(LLONG_MIN));
-}
-
-TEST_CASE("should bind SQL_C_SBIGINT LLONG_MAX to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLBIGINT param = LLONG_MAX;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SBIGINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == std::to_string(LLONG_MAX));
-}
-
-TEST_CASE("should bind SQL_C_UBIGINT ULLONG_MAX to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLUBIGINT param = ULLONG_MAX;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_UBIGINT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == std::to_string(ULLONG_MAX));
-}
-
-TEST_CASE("should bind SQL_C_DOUBLE negative zero to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLDOUBLE param = -0.0;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DOUBLE, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "0");
-}
-
-static std::string to_lower(const std::string& s) {
-  std::string lower = s;
-  std::transform(lower.begin(), lower.end(), lower.begin(), [](unsigned char c) { return std::tolower(c); });
-  return lower;
-}
-
-static bool is_nan_str(const std::string& s) { return to_lower(s) == "nan"; }
-
-static bool is_positive_infinity_str(const std::string& s) {
-  auto l = to_lower(s);
-  return l == "inf" || l == "infinity";
-}
-
-static bool is_negative_infinity_str(const std::string& s) {
-  auto l = to_lower(s);
-  return l == "-inf" || l == "-infinity";
-}
-
-TEST_CASE("should bind SQL_C_DOUBLE NaN to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLDOUBLE param = std::numeric_limits<double>::quiet_NaN();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DOUBLE, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_nan_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_DOUBLE positive infinity to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLDOUBLE param = std::numeric_limits<double>::infinity();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DOUBLE, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_positive_infinity_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_DOUBLE negative infinity to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLDOUBLE param = -std::numeric_limits<double>::infinity();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DOUBLE, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_negative_infinity_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_FLOAT NaN to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLREAL param = std::numeric_limits<float>::quiet_NaN();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_FLOAT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_nan_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_FLOAT positive infinity to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLREAL param = std::numeric_limits<float>::infinity();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_FLOAT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_positive_infinity_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_FLOAT negative infinity to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLREAL param = -std::numeric_limits<float>::infinity();
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_FLOAT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(is_negative_infinity_str(get_data<SQL_C_CHAR>(stmt, 1)));
-}
-
-TEST_CASE("should bind SQL_C_BIT value > 1 to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLCHAR param = 42;
-  SQLLEN indicator = sizeof(param);
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_BIT, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "1");
-}
-
-TEST_CASE("should bind SQL_C_DEFAULT to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  Connection conn;
-  auto stmt = conn.createStatement();
-  char param[] = "hello";
-  SQLLEN indicator = SQL_NTS;
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_DEFAULT, SQL_VARCHAR, 100, 0, param,
-                                   sizeof(param), &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "hello");
-}
-
-// =============================================================================
-// C type → other string SQL types (CHAR, WCHAR, LONGVARCHAR, WVARCHAR, WLONGVARCHAR)
-// Verify that all string SQL types correctly route through the same conversion.
-// =============================================================================
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_CHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_CHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_LONGVARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_LONGVARCHAR, 100, 0, &param,
-                                   0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_WCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_WCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_WVARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_WVARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind SQL_C_SLONG to SQL_WLONGVARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQLINTEGER param = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_SLONG, SQL_WLONGVARCHAR, 100, 0, &param,
-                                   0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-// =============================================================================
-// C structured type → VARCHAR conversions
-// =============================================================================
-
-TEST_CASE("should bind SQL_C_TYPE_TIMESTAMP to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_TIMESTAMP_STRUCT param = {};
-  param.year = 2024;
-  param.month = 1;
-  param.day = 15;
-  param.hour = 10;
-  param.minute = 30;
-  param.second = 45;
-  param.fraction = 0;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_TYPE_TIMESTAMP, SQL_VARCHAR, 100, 0,
-                                   &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "2024-01-15 10:30:45");
-}
-
-TEST_CASE("should bind SQL_C_TYPE_TIMESTAMP with fraction to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_TIMESTAMP_STRUCT param = {};
-  param.year = 2024;
-  param.month = 6;
-  param.day = 15;
-  param.hour = 12;
-  param.minute = 0;
-  param.second = 0;
-  param.fraction = 123456789;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_TYPE_TIMESTAMP, SQL_VARCHAR, 100, 0,
-                                   &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "2024-06-15 12:00:00.123456789");
-}
-
-TEST_CASE("should bind SQL_C_TYPE_DATE to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_DATE_STRUCT param = {};
-  param.year = 2024;
-  param.month = 12;
-  param.day = 25;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_TYPE_DATE, SQL_VARCHAR, 100, 0, &param,
-                                   0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "2024-12-25");
-}
-
-TEST_CASE("should bind SQL_C_TYPE_TIME to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_TIME_STRUCT param = {};
-  param.hour = 14;
-  param.minute = 30;
-  param.second = 59;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_TYPE_TIME, SQL_VARCHAR, 100, 0, &param,
-                                   0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "14:30:59");
-}
-
-TEST_CASE("should bind SQL_C_NUMERIC to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_NUMERIC_STRUCT param = {};
-  param.precision = 10;
-  param.scale = 0;
-  param.sign = 1;
-  memset(param.val, 0, SQL_MAX_NUMERIC_LEN);
-  param.val[0] = 42;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_NUMERIC, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "42");
-}
-
-TEST_CASE("should bind negative SQL_C_NUMERIC with scale to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_NUMERIC_STRUCT param = {};
-  param.precision = 10;
-  param.scale = 2;
-  param.sign = 0;
-  memset(param.val, 0, SQL_MAX_NUMERIC_LEN);
-  // 12345 in little-endian: 0x39, 0x30
-  param.val[0] = 0x39;
-  param.val[1] = 0x30;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_NUMERIC, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  auto result = get_data<SQL_C_CHAR>(stmt, 1);
-  NEW_DRIVER_ONLY("BD#33") { CHECK(result == "-123.45"); }
-  OLD_DRIVER_ONLY("BD#33") { CHECK(result == "-12345"); }
-}
-
-TEST_CASE("should bind SQL_C_NUMERIC with negative scale to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  SQL_NUMERIC_STRUCT param = {};
-  param.precision = 10;
-  param.scale = static_cast<SQLSCHAR>(-2);
-  param.sign = 1;
-  memset(param.val, 0, SQL_MAX_NUMERIC_LEN);
-  param.val[0] = 123;
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret =
-      SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_NUMERIC, SQL_VARCHAR, 100, 0, &param, 0, &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  REQUIRE_ODBC(ret, stmt);
-  ret = SQLFetch(stmt.getHandle());
-  REQUIRE_ODBC(ret, stmt);
-  // Then the result should be the expected string
-  auto result = get_data<SQL_C_CHAR>(stmt, 1);
-  NEW_DRIVER_ONLY("BD#33") { CHECK(result == "12300"); }
-  OLD_DRIVER_ONLY("BD#33") { CHECK(result == "123"); }
-}
-
-TEST_CASE("should bind SQL_C_BINARY to SQL_VARCHAR.", "[query][bind_parameter][c_to_varchar]") {
-  // Given Snowflake client is logged in
-  Connection conn;
-  auto stmt = conn.createStatement();
-  unsigned char param[] = {0xDE, 0xAD, 0xBE, 0xEF};
-  SQLLEN indicator = sizeof(param);
-  // When the C type value is bound as a string SQL type and SELECT ? is executed
-  SQLRETURN ret = SQLBindParameter(stmt.getHandle(), 1, SQL_PARAM_INPUT, SQL_C_BINARY, SQL_VARCHAR, 100, 0, &param,
-                                   sizeof(param), &indicator);
-  REQUIRE_ODBC_SUCCESS(ret, stmt);
-  ret = SQLExecDirect(stmt.getHandle(), sqlchar("SELECT ? AS val"), SQL_NTS);
-  NEW_DRIVER_ONLY("BD#34") {
-    REQUIRE_ODBC(ret, stmt);
-    ret = SQLFetch(stmt.getHandle());
-    REQUIRE_ODBC(ret, stmt);
-    // Then the result should be the expected string
-    CHECK(get_data<SQL_C_CHAR>(stmt, 1) == "deadbeef");
-  }
-  OLD_DRIVER_ONLY("BD#34") { CHECK(ret == SQL_ERROR); }
 }
